added common header component
