+++ src/App.vue
@@ -1,6 +1,6 @@
 <template>
+  <Header />
   <main font-sans p="x-4 y-10" text="center gray-700 dark:gray-200">
     <router-view />
-    <Footer />
   </main>
 </template>
